Code changes, add coins button
Added all the code for the "Add New Coin" buttons. when the Plus sign is selected next to a coin it adds 1 to the quantity of coins by changing the label text associated with the coin. Each block of coin add code also calls the addCoins() and totalCredits() functions I defined in the previous 2 Commits of GitHUB.