Changes To Code
Linked up the button code to the actions of the click event.
Added the reset button code which resets all the labels and text boxes to 0.
Also made the click event on the actual coin pictures work so you can either use the add button or the coin itself to change coin quantity.